+++ src/server-manager.cpp
@@ -56,7 +56,7 @@
 #pragma GCC diagnostic pop
 
 static constexpr uint32_t manager_version_major = 0;
-static constexpr uint32_t manager_version_minor = 2;
+static constexpr uint32_t manager_version_minor = 3;
 static constexpr uint32_t manager_version_patch = 0;
 static constexpr const char *manager_version_pre_release = "";
 static constexpr const char *manager_version_build = "";