+++ src/test-server-manager.cpp
@@ -665,12 +665,12 @@ void assert_bool(bool value, const std::string &msg, const std::string &file, in
 #define ASSERT_STATUS_CODE(s, f, v) assert_status_code(s, f, v, __FILE__, __LINE__)
 
 static void test_get_version(const std::function<void(const std::string &title, test_function f)> &test) {
-    test("The rollup-machine-manager server version should be 0.2.x", [](ServerManagerClient &manager) {
+    test("The rollup-machine-manager server version should be 0.3.x", [](ServerManagerClient &manager) {
         Versioning::GetVersionResponse response;
         Status status = manager.get_version(response);
         ASSERT_STATUS(status, "GetVersion", true);
         ASSERT((response.version().major() == 0), "Version Major should be 0");
-        ASSERT((response.version().minor() == 2), "Version Minor should be 2");
+        ASSERT((response.version().minor() == 3), "Version Minor should be 3");
     });
 }